requirements.txt
# QC Tool Dependencies
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
pyathena>=3.0.0
python-dotenv>=1.0.0
xlsxwriter>=3.1.0
